remove tabs navigation in dashboard and move settings back
