+++ src/app/account-management/account/account.component.html
@@ -1,10 +1,10 @@
-<!-- <div id="account">
+<div id="account">
     <div class="banner">
         <h3>Account</h3>
     </div>
-</div> -->
+</div>
 
-<div class="wrapper m-0">
+<div class="wrapper">
     <h4>Corporation Information</h4>
     <br>
     <div class="row">
