+++ src/app/account-management/facility/facility.component.html
@@ -1,10 +1,10 @@
-<!-- <div id="facility">
+<div id="facility">
     <div class="banner">
         <h3>Facility Information</h3>
     </div>
-</div> -->
+</div>
 
-<div class="wrapper m-0">
+<div class="wrapper">
     <h4>Facilty Information</h4>
     <br>
     <div class="row">
