+++ src/app/dashboard/account-overview/account-overview.component.ts
@@ -7,7 +7,7 @@ import { IdbFacility } from 'src/app/models/idb';
 @Component({
   selector: 'app-account-overview',
   templateUrl: './account-overview.component.html',
-  styleUrls: ['./account-overview.component.css', '../../dashboard.component.css']
+  styleUrls: ['./account-overview.component.css', '../dashboard.component.css']
 })
 export class AccountOverviewComponent implements OnInit {
