+++ src/app/dashboard/account-overview/energy-use-heat-map/energy-use-heat-map.component.spec.ts
