+++ src/app/dashboard/account-overview/energy-use-stacked-bar-chart/energy-use-stacked-bar-chart.component.ts
@@ -3,7 +3,6 @@ import { PlotlyService } from 'angular-plotly.js';
 import { Subscription } from 'rxjs';
 import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db.service';
 import { IdbFacility, IdbUtilityMeter, IdbUtilityMeterData } from 'src/app/models/idb';
-import { AccountSummaryService } from '../../account-summary.service';
 import * as _ from 'lodash';
 import { UtilityMeterdbService } from 'src/app/indexedDB/utilityMeter-db.service';
 import { FacilitydbService } from 'src/app/indexedDB/facility-db.service';
