--- src/app/dashboard/account-summary/account-summary.component.html
@@ -1,19 +1,0 @@
-<div class="dashboard-nav">
-  <ul>
-    <!-- <li >Summary</li>
-      <li [disabled] class="active">Utilities</li>
-      <li>Water</li> -->
-    <li routerLink="/account-summary/overview" [routerLinkActive]="['active']">Summary</li>
-    <!-- <li routerLink="utilities" [routerLinkActive]="['active']">All Utilities</li> -->
-    <!-- <li><a routerLink="meter-groups" [routerLinkActive]="['active']">Meter Grouping</a></li>
-      <li><a routerLink="predictors" [routerLinkActive]="['active']">Predictors</a></li>
-      <li><a routerLink="visualization" [routerLinkActive]="['active']">Data Visualization</a></li> -->
-  </ul>
-  <!-- <button class="btn btn-outline item-right">Generate Report</button> -->
-  <ul class="item-right">
-    <li routerLink="/account-summary/management" [routerLinkActive]="['active']">Manage</li>
-  </ul>
-</div>
-<div class="wrapper main-content">
-  <router-outlet></router-outlet>
-</div>
