--- src/app/dashboard/account-summary/account-summary.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AccountSummaryService } from './account-summary.service';
-
-describe('AccountSummaryService', () => {
-  let service: AccountSummaryService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(AccountSummaryService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
