+++ src/app/dashboard/dashboard.component.html
@@ -9,7 +9,9 @@
   <h3 *ngIf="selectedAccount && !facilityDashboard">
     {{selectedAccount.name}}
   </h3>
-  <button class="m-0 btn btn-outline item-right">Generate Report</button> 
+  <button class="m-0 btn btn-outline item-right">Generate Report</button>
 </div>
-<router-outlet></router-outlet>
+<div class="wrapper main-content">
+  <router-outlet></router-outlet>
+</div>
 <div class="measurFooter"></div>
