+++ src/app/dashboard/facility-overview/facility-overview.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'app-facility-overview',
   templateUrl: './facility-overview.component.html',
-  styleUrls: ['./facility-overview.component.css', '../../dashboard.component.css']
+  styleUrls: ['./facility-overview.component.css', '../dashboard.component.css']
 })
 export class FacilityOverviewComponent implements OnInit {
