--- src/app/dashboard/facility-summary/facility-summary.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-facility-summary',
-  templateUrl: './facility-summary.component.html',
-  styleUrls: ['./facility-summary.component.css', '../dashboard.component.css']
-})
-export class FacilitySummaryComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
