+++ src/app/shared/header/header.component.html
@@ -27,7 +27,7 @@
                         <p class="small">Switch Facility <i class="fa fa-random item-right"></i></p>
                         <p (click)="switchFacility(facility)" *ngFor="let facility of facilityList">
                             {{facility.name}}
-                            <i class="fa fa-cog item-right" routerLink="/facility-summary/management"
+                            <i class="fa fa-cog item-right" routerLink="/facility-management"
                                 title="Manage Facility Settings"></i>
                         </p>
                     </div>
@@ -40,7 +40,7 @@
                 <i class="fa fa-ellipsis-v" (click)="toggleAccountMenu()"></i>
                 <div class="dropdown" [class.open]="accountMenu == true">
                     <div class="arrow-up"></div>
-                    <div class="currentAccount" routerLink="/account-summary/management">
+                    <div class="currentAccount" routerLink="/account-management">
                         <i class="fa fa-user-circle-o item-left"></i>
                         <p>{{activeAccount.name}}</p>
                         <span class="small">Manage Account</span>
